Finally got the bloody FBO working now (to a degree), some weird problems happen to the render if i draw springs for some odd reason, so i'll check that out later, i can draw to a texture which is GOOD. I'm currently limiting fps (it's suppose to be 60, but it's coming out to 63) it's sorta working, but if we try to load the waypoints, the way limiting fps is done is the game thinks the game is still running during the load, so once it finished loading all the waypoints, the game tries to catch up to the fps it should be at. This causes some EXTREME laggy moment of super speed until the game catches back up. I also optimized drawing (hopefully), i'm currently using vertex/normal/texture/whatever pointers for opengl, so this is pretty darn fast for drawing (i can optimize it further with VBO's). Also got the texture to work properly sorta, one side (the left) is still borked.
Some minor changes here and there. There's a bit of a lag spike when we
hit a power up (it's hacked), but we can worry about that later.

I hope that I can get a car model AND shadows working (cross my fingers).

I might also try to do a hacked GUI together just so everything can be
loaded at the start of the game, and once some flag gets set, the game
will "start", this is so loading the waypoints on the fly is better? ish.

Giant wall of text is giant.